@@ -1,9 +1,9 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/rendering.dart';
 import 'dart:ui' as ui;
 
 // ignore: must_be_immutable
-class HorizontalNumberPicker extends StatefulWidget {
+class VerticalNumberPicker extends StatefulWidget {
   final int initialValue;
   final int minValue;
   final int maxValue;
@@ -21,15 +21,15 @@ class HorizontalNumberPicker extends StatefulWidget {
   ///一大格中有多少个小格
   final int subGridCountPerGrid;
 
-  ///大格的宽度
-  late int gridWidth;
+  ///大格的高度
+  late int gridHeight;
 
-  ///每一小格的宽度
-  final int subGridWidth;
+  ///每一小格的高度
+  final int subGridHeight;
 
   late int listViewItemCount;
 
-  late double paddingItemWidth;
+  late double paddingItemHeight;
 
   final void Function(int) onSelectedChanged;
 
@@ -45,7 +45,7 @@ class HorizontalNumberPicker extends StatefulWidget {
   ///刻度文字颜色
   final Color scaleTextColor;
 
-  HorizontalNumberPicker({
+  VerticalNumberPicker({
     super.key,
     this.initialValue = 500,
     this.minValue = 100,
@@ -54,12 +54,12 @@ class HorizontalNumberPicker extends StatefulWidget {
     this.widgetWidth = 200,
     this.widgetHeight = 60,
     this.subGridCountPerGrid = 10,
-    this.subGridWidth = 8,
+    this.subGridHeight = 8,
     required this.onSelectedChanged,
     this.scaleTransformer,
-    this.scaleColor = const Color(0xFFE9E9E9),
-    this.indicatorColor = const Color(0xFF3995FF),
-    this.scaleTextColor = const Color(0xFF8E99A0),
+    this.scaleColor = const Color(0xff000000), // const Color(0xffD7DCE7),
+    this.indicatorColor = const Color(0xff036AFF),
+    this.scaleTextColor = const Color(0x803B4664),
   }) : super() {
     if (subGridCountPerGrid % 2 != 0) {
       throw Exception("subGridCountPerGrid必须是偶数");
@@ -76,30 +76,28 @@ class HorizontalNumberPicker extends StatefulWidget {
     //第一个grid和最后一个grid都只会展示一半数量的subGrid，因此gridCount需要+1
     gridCount = totalSubGridCount ~/ subGridCountPerGrid + 1;
 
-    gridWidth = subGridWidth * subGridCountPerGrid;
+    gridHeight = subGridHeight * subGridCountPerGrid;
 
     //每个grid都是listView的一个item
     //除此之外，在第一个grid之前和最后一个grid之后，还需要各填充一个空白item，
     //这样第一个item和最后一个item才能滚动到屏幕中间。
     listViewItemCount = gridCount + 2;
 
-    //空白item的宽度
-    paddingItemWidth = widgetWidth / 2 - gridWidth / 2;
+    //空白item的高度
+    paddingItemHeight = (widgetHeight - gridHeight) / 2;
 
-    if (scaleTransformer == null) {
-      scaleTransformer = (value) {
-        return value.toString();
-      };
-    }
+    scaleTransformer ??= (value) {
+      return value.toString();
+    };
   }
 
   @override
   State<StatefulWidget> createState() {
-    return HorizontalNumberPickerState();
+    return VerticalNumberPickerState();
   }
 }
 
-class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
+class VerticalNumberPickerState extends State<VerticalNumberPicker> {
   late ScrollController _scrollController;
 
   @override
@@ -108,13 +106,14 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
       //计算初始偏移量
       initialScrollOffset: (widget.initialValue - widget.minValue) /
           widget.step *
-          widget.subGridWidth,
+          widget.subGridHeight,
     );
     super.initState();
   }
 
   ///处理state的复用
-  void didUpdateWidget(HorizontalNumberPicker oldWidget) {
+  @override
+  void didUpdateWidget(VerticalNumberPicker oldWidget) {
     super.didUpdateWidget(oldWidget);
 
     _scrollController.dispose();
@@ -122,7 +121,7 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
       //计算初始偏移量
       initialScrollOffset: (widget.initialValue - widget.minValue) /
           widget.step *
-          widget.subGridWidth,
+          widget.subGridHeight,
     );
   }
 
@@ -132,7 +131,7 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
       width: widget.widgetWidth.toDouble(),
       height: widget.widgetHeight.toDouble(),
       child: Stack(
-        alignment: Alignment.topCenter,
+        alignment: Alignment.center,
         children: <Widget>[
           NotificationListener(
             onNotification: _onNotification,
@@ -140,14 +139,14 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
               physics: const ClampingScrollPhysics(),
               padding: const EdgeInsets.all(0),
               controller: _scrollController,
-              scrollDirection: Axis.horizontal,
+              scrollDirection: Axis.vertical,
               itemCount: widget.listViewItemCount,
               itemBuilder: (BuildContext context, int index) {
                 //首尾空白元素
                 if (index == 0 || index == widget.listViewItemCount - 1) {
                   return SizedBox(
-                    width: widget.paddingItemWidth,
-                    height: 0,
+                    width: 0,
+                    height: widget.paddingItemHeight,
                   );
                   //普通元素
                 } else {
@@ -163,15 +162,35 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
                     type = 1;
                   }
 
-                  return NumberPickerItem(
-                    subGridCount: widget.subGridCountPerGrid,
-                    subGridWidth: widget.subGridWidth,
-                    itemHeight: widget.widgetHeight,
-                    valueStr: widget.scaleTransformer!(widget.minValue +
-                        (index - 1) * widget.subGridCountPerGrid * widget.step),
-                    type: type,
-                    scaleColor: widget.scaleColor,
-                    scaleTextColor: widget.scaleTextColor,
+                  // return NumberPickerItem(
+                  //   subGridCount: widget.subGridCountPerGrid,
+                  //   subGridHeight: widget.subGridHeight,
+                  //   itemWidth: widget.widgetWidth,
+                  //   valueStr: widget.scaleTransformer!(widget.minValue +
+                  //       (index - 1) * widget.subGridCountPerGrid * widget.step),
+                  //   type: type,
+                  //   scaleColor: widget.scaleColor,
+                  //   scaleTextColor: widget.scaleTextColor,
+                  // );
+                  return GestureDetector(
+                    onTap: () {
+                      select(widget.minValue +
+                          (index - 1) *
+                              widget.subGridCountPerGrid *
+                              widget.step);
+                    },
+                    child: NumberPickerItem(
+                      subGridCount: widget.subGridCountPerGrid,
+                      subGridHeight: widget.subGridHeight,
+                      itemWidth: widget.widgetWidth,
+                      valueStr: widget.scaleTransformer!(widget.minValue +
+                          (index - 1) *
+                              widget.subGridCountPerGrid *
+                              widget.step),
+                      type: type,
+                      scaleColor: widget.scaleColor,
+                      scaleTextColor: widget.scaleTextColor,
+                    ),
                   );
                 }
               },
@@ -179,8 +198,8 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
           ),
           //指示器
           Container(
-            width: 2,
-            height: widget.widgetHeight * 0.55,
+            width: widget.widgetWidth * 0.55,
+            height: 2,
             color: widget.indicatorColor,
           ),
         ],
@@ -193,7 +212,7 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
     if (notification is ScrollNotification) {
       //距离widget中间最近的刻度值
       int centerValue =
-          (notification.metrics.pixels / widget.subGridWidth).round() *
+          (notification.metrics.pixels / widget.subGridHeight).round() *
                   widget.step +
               widget.minValue;
 
@@ -224,8 +243,8 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
   ///选中值
   select(int valueToSelect) {
     _scrollController.animateTo(
-      (valueToSelect - widget.minValue) / widget.step * widget.subGridWidth,
-      duration: Duration(milliseconds: 200),
+      (valueToSelect - widget.minValue) / widget.step * widget.subGridHeight,
+      duration: const Duration(milliseconds: 200),
       curve: Curves.decelerate,
     );
   }
@@ -236,8 +255,8 @@ class HorizontalNumberPickerState extends State<HorizontalNumberPicker> {
 ///每个item中间为长刻度，并在下方显示数值。两边都是短刻度
 class NumberPickerItem extends StatelessWidget {
   final int subGridCount;
-  final int subGridWidth;
-  final int itemHeight;
+  final int subGridHeight;
+  final int itemWidth;
   final String valueStr;
 
   //0:列表首item 1:中间item 2:尾item
@@ -249,8 +268,8 @@ class NumberPickerItem extends StatelessWidget {
   const NumberPickerItem({
     super.key,
     required this.subGridCount,
-    required this.subGridWidth,
-    required this.itemHeight,
+    required this.subGridHeight,
+    required this.itemWidth,
     required this.valueStr,
     required this.type,
     required this.scaleColor,
@@ -259,19 +278,19 @@ class NumberPickerItem extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    double itemWidth = (subGridWidth * subGridCount).toDouble();
-    double itemHeight = this.itemHeight.toDouble();
+    double itemWidth = this.itemWidth.toDouble();
+    double itemHeight = (subGridHeight * subGridCount).toDouble();
 
     return CustomPaint(
       size: Size(itemWidth, itemHeight),
-      painter: MyPainter(this.subGridWidth, this.valueStr, this.type,
-          this.scaleColor, this.scaleTextColor),
+      painter:
+          MyPainter(subGridHeight, valueStr, type, scaleColor, scaleTextColor),
     );
   }
 }
 
 class MyPainter extends CustomPainter {
-  final int subGridWidth;
+  final int subGridHeight;
 
   final String valueStr;
 
@@ -284,10 +303,10 @@ class MyPainter extends CustomPainter {
 
   late Paint _linePaint;
 
-  final double _lineWidth = 2;
+  final double _lineHeight = 2;
 
   MyPainter(
-    this.subGridWidth,
+    this.subGridHeight,
     this.valueStr,
     this.type,
     this.scaleColor,
@@ -296,7 +315,7 @@ class MyPainter extends CustomPainter {
     _linePaint = Paint()
       ..isAntiAlias = true
       ..style = PaintingStyle.stroke
-      ..strokeWidth = _lineWidth
+      ..strokeWidth = _lineHeight
       ..color = scaleColor;
   }
 
@@ -307,46 +326,58 @@ class MyPainter extends CustomPainter {
   }
 
   void drawLine(Canvas canvas, Size size) {
-    double startX, endX;
+    double startY, endY;
     switch (type) {
-      case 0: //首元素只绘制右半部分
-        startX = size.width / 2;
-        endX = size.width;
+      case 0: //首元素只绘制下半部分
+        startY = size.height / 2;
+        endY = size.height;
         break;
-      case 2: //尾元素只绘制左半部分
-        startX = 0;
-        endX = size.width / 2;
+      case 2: //尾元素只绘制上半部分
+        startY = 0;
+        endY = size.height / 2;
         break;
       default: //中间元素全部绘制
-        startX = 0;
-        endX = size.width;
+        startY = 0;
+        endY = size.height;
     }
 
-    //绘制横线
-    canvas.drawLine(Offset(startX, 0 + _lineWidth / 2),
-        Offset(endX, 0 + _lineWidth / 2), _linePaint);
+    double largeLine = size.width * 0.383;
+    double smallLine = size.width * 0.183;
+    double offset = (largeLine - smallLine) / 2;
 
     //绘制竖线
-    for (double x = startX; x <= endX; x += subGridWidth) {
-      if (x == size.width / 2) {
+    canvas.drawLine(
+        Offset(largeLine / 2, startY), Offset(largeLine / 2, endY), _linePaint);
+
+    //绘制横线
+    for (double y = startY; y <= endY; y += subGridHeight) {
+      // if (x == size.width / 2) {
+      //   //中间为长刻度
+      //   canvas.drawLine(
+      //       Offset(x, 0), Offset(x, size.height * 0.183), _linePaint);
+      // } else {
+      //   //其他为短刻度
+      //   canvas.drawLine(
+      //       Offset(x, 0), Offset(x, size.height * 0.383), _linePaint);
+      // }
+      if (y == size.height / 2) {
         //中间为长刻度
-        canvas.drawLine(
-            Offset(x, 0), Offset(x, size.height * 3 / 8), _linePaint);
+        canvas.drawLine(Offset(0, y), Offset(largeLine, y), _linePaint);
       } else {
         //其他为短刻度
-        canvas.drawLine(Offset(x, 0), Offset(x, size.height / 4), _linePaint);
+        canvas.drawLine(
+            Offset(offset, y), Offset(offset + smallLine, y), _linePaint);
       }
     }
   }
 
   void drawText(Canvas canvas, Size size) {
     //文字水平方向居中对齐，竖直方向底对齐
-    ui.Paragraph p = _buildText(valueStr, size.width);
+    ui.Paragraph p = _buildText(valueStr, size.height);
     //获得文字的宽高
     double halfWidth = p.minIntrinsicWidth / 2;
-    // double halfHeight = p.height / 2;
     canvas.drawParagraph(
-        p, Offset(size.width / 2 - halfWidth, size.height - p.height));
+        p, Offset(size.width / 2 - halfWidth, (size.height - p.height) / 2));
   }
 
   ui.Paragraph _buildText(String content, double maxWidth) {
@@ -354,9 +385,9 @@ class MyPainter extends CustomPainter {
         ui.ParagraphBuilder(ui.ParagraphStyle());
     paragraphBuilder.pushStyle(
       ui.TextStyle(
-        fontSize: 14,
-        color: this.scaleTextColor,
-        //fontFamily: "Montserrat",
+        color: scaleTextColor,
+        fontSize: 16,
+        fontWeight: FontWeight.w400,
       ),
     );
     paragraphBuilder.addText(content);
